Reorganize tests to be per-model.
