+++ Tests/Codecs/VocosTests.swift
@@ -1,0 +1,253 @@
+import Testing
+import MLX
+import Foundation
+
+@testable import MLXAudioCore
+@testable import MLXAudioTTS
+@testable import MLXAudioCodecs
+@testable import MLXAudioLID
+
+struct SharedDSPTests {
+
+    @Test func hammingWindowSupportsPeriodicAndSymmetricVariants() {
+        let periodic = hammingWindow(size: 4).asArray(Float.self)
+        let symmetric = hammingWindow(size: 4, periodic: false).asArray(Float.self)
+
+        #expect(periodic.count == 4)
+        #expect(symmetric.count == 4)
+
+        #expect(abs(periodic[0] - 0.08) < 1e-3)
+        #expect(abs(periodic[1] - 0.54) < 1e-3)
+        #expect(abs(periodic[3] - 0.54) < 1e-3)
+
+        #expect(abs(symmetric[0] - 0.08) < 1e-3)
+        #expect(abs(symmetric[3] - 0.08) < 1e-3)
+        #expect(abs(symmetric[1] - symmetric[2]) < 1e-3)
+    }
+
+    @Test func powerToDBAppliesTopDBClipping() {
+        let spectrogram = MLXArray([Float(1e-10), Float(1e-5), Float(1.0)])
+        let clipped = powerToDB(spectrogram, topDB: 80).asArray(Float.self)
+
+        #expect(abs(clipped[0] + 80) < 1e-2)
+        #expect(abs(clipped[1] + 50) < 1e-2)
+        #expect(abs(clipped[2]) < 1e-3)
+    }
+}
+
+
+// MARK: - Vocos Tests
+// Run Vocos tests with:  xcodebuild test \
+// -scheme MLXAudio-Package \
+// -destination 'platform=macOS' \
+// -only-testing:MLXAudioTests/VocosTests \
+// 2>&1 | grep -E "(Suite.*started|Test test.*started|passed after|failed after|TEST SUCCEEDED|TEST FAILED|Suite.*passed|Test run)"
+
+
+struct VocosTests {
+
+    @Test func testConvNeXtBlock() throws {
+        // Test basic ConvNeXtBlock forward pass
+        let dim = 64
+        let intermediateDim = 192
+        let block = ConvNeXtBlock(
+            dim: dim,
+            intermediateDim: intermediateDim,
+            layerScaleInitValue: 0.125,
+            dwKernelSize: 7
+        )
+
+        // Input shape: (batch, length, channels)
+        let input = MLXRandom.normal([1, 100, dim])
+        let output = block(input)
+
+        // Output should have same shape as input (residual connection)
+        #expect(output.shape == input.shape)
+        print("ConvNeXtBlock output shape: \(output.shape)")
+    }
+
+    @Test func testConvNeXtBlockWithAdaNorm() throws {
+        // Test ConvNeXtBlock with adaptive normalization
+        let dim = 64
+        let intermediateDim = 192
+        let numEmbeddings = 4
+
+        let block = ConvNeXtBlock(
+            dim: dim,
+            intermediateDim: intermediateDim,
+            layerScaleInitValue: 0.125,
+            adanormNumEmbeddings: numEmbeddings,
+            dwKernelSize: 7
+        )
+
+        // Input shape: (batch, length, channels)
+        let input = MLXRandom.normal([1, 100, dim])
+        let condEmbedding = MLXRandom.normal([1, numEmbeddings])
+
+        let output = block(input, condEmbeddingId: condEmbedding)
+
+        // Output should have same shape as input
+        #expect(output.shape == input.shape)
+        print("ConvNeXtBlock with AdaNorm output shape: \(output.shape)")
+    }
+
+    @Test func testVocosBackbone() throws {
+        // Test VocosBackbone forward pass
+        let inputChannels = 100
+        let dim = 512
+        let intermediateDim = 1536
+        let numLayers = 8
+
+        let backbone = VocosBackbone(
+            inputChannels: inputChannels,
+            dim: dim,
+            intermediateDim: intermediateDim,
+            numLayers: numLayers
+        )
+
+        // Input shape: (batch, length, input_channels)
+        let input = MLXRandom.normal([1, 50, inputChannels])
+        let output = backbone(input)
+
+        // Output should have shape (batch, length, dim)
+        #expect(output.shape[0] == 1)
+        #expect(output.shape[1] == 50)
+        #expect(output.shape[2] == dim)
+        print("VocosBackbone output shape: \(output.shape)")
+    }
+
+    @Test func testVocosBackboneWithAdaNorm() throws {
+        // Test VocosBackbone with adaptive normalization
+        let inputChannels = 100
+        let dim = 256
+        let intermediateDim = 768
+        let numLayers = 4
+        let numEmbeddings = 4
+
+        let backbone = VocosBackbone(
+            inputChannels: inputChannels,
+            dim: dim,
+            intermediateDim: intermediateDim,
+            numLayers: numLayers,
+            adanormNumEmbeddings: numEmbeddings
+        )
+
+        // Input shape: (batch, length, input_channels)
+        let input = MLXRandom.normal([1, 50, inputChannels])
+        let bandwidthId = MLXRandom.normal([1, numEmbeddings])
+
+        let output = backbone(input, bandwidthId: bandwidthId)
+
+        // Output should have shape (batch, length, dim)
+        #expect(output.shape[0] == 1)
+        #expect(output.shape[1] == 50)
+        #expect(output.shape[2] == dim)
+        print("VocosBackbone with AdaNorm output shape: \(output.shape)")
+    }
+
+    @Test func testISTFTHead() throws {
+        // Test ISTFTHead forward pass
+        let dim = 512
+        let nFft = 1024
+        let hopLength = 256
+
+        let head = MLXAudioCodecs.ISTFTHead(dim: dim, nFft: nFft, hopLength: hopLength)
+
+        // Input shape: (batch, length, dim)
+        let numFrames = 100
+        let input = MLXRandom.normal([1, numFrames, dim])
+
+        let output = head(input)
+
+        // Output should be audio waveform
+        // Expected length: approximately (numFrames - 1) * hopLength after trimming
+        #expect(output.ndim == 1 || output.ndim == 2)
+        print("ISTFTHead output shape: \(output.shape)")
+    }
+
+    @Test func testAdaLayerNorm() throws {
+        // Test AdaLayerNorm
+        let numEmbeddings = 4
+        let embeddingDim = 256
+
+        let adaNorm = AdaLayerNorm(
+            numEmbeddings: numEmbeddings,
+            embeddingDim: embeddingDim
+        )
+
+        // Input shape: (batch, length, dim)
+        let input = MLXRandom.normal([2, 50, embeddingDim])
+        let condEmbedding = MLXRandom.normal([2, numEmbeddings])
+
+        let output = adaNorm(input, condEmbedding: condEmbedding)
+
+        // Output should have same shape as input
+        #expect(output.shape == input.shape)
+        print("AdaLayerNorm output shape: \(output.shape)")
+    }
+
+    @Test func testVocosModel() throws {
+        // Test full Vocos model
+        let inputChannels = 100
+        let dim = 256
+        let intermediateDim = 768
+        let numLayers = 4
+        let nFft = 1024
+        let hopLength = 256
+
+        let backbone = VocosBackbone(
+            inputChannels: inputChannels,
+            dim: dim,
+            intermediateDim: intermediateDim,
+            numLayers: numLayers
+        )
+
+        let head = MLXAudioCodecs.ISTFTHead(dim: dim, nFft: nFft, hopLength: hopLength)
+
+        let vocos = Vocos(backbone: backbone, head: head)
+
+        // Input shape: (batch, length, input_channels)
+        let numFrames = 50
+        let input = MLXRandom.normal([1, numFrames, inputChannels])
+
+        let output = vocos(input)
+
+        // Output should be audio waveform
+        print("Vocos output shape: \(output.shape)")
+        #expect(output.shape.count >= 1)
+    }
+
+    @Test func testVocosDecodeWithBandwidthId() throws {
+        // Test Vocos decode with bandwidth conditioning
+        let inputChannels = 128
+        let dim = 256
+        let intermediateDim = 768
+        let numLayers = 4
+        let numEmbeddings = 4
+        let nFft = 1024
+        let hopLength = 256
+
+        let backbone = VocosBackbone(
+            inputChannels: inputChannels,
+            dim: dim,
+            intermediateDim: intermediateDim,
+            numLayers: numLayers,
+            adanormNumEmbeddings: numEmbeddings
+        )
+
+        let head = MLXAudioCodecs.ISTFTHead(dim: dim, nFft: nFft, hopLength: hopLength)
+
+        let vocos = Vocos(backbone: backbone, head: head)
+
+        // Input shape: (batch, length, input_channels)
+        let numFrames = 50
+        let input = MLXRandom.normal([1, numFrames, inputChannels])
+        let bandwidthId = MLXRandom.normal([1, numEmbeddings])
+
+        let output = vocos.decode(input, bandwidthId: bandwidthId)
+
+        // Output should be audio waveform
+        print("Vocos decode with bandwidthId output shape: \(output.shape)")
+        #expect(output.shape.count >= 1)
+    }
+}
