--- Tests/MLXAudioTTSTests.swift
@@ -1,42 +1,0 @@
-//
-//  MLXAudioTTSTests.swift
-//  MLXAudioTests
-//
-//  Created by Prince Canuma on 31/12/2025.
-//
-
-import Testing
-import MLX
-import MLXLMCommon
-import Foundation
-
-@testable import MLXAudioCore
-@testable import MLXAudioTTS
-@testable import MLXAudioCodecs
-
-
-// MARK: - Text Cleaning Unit Tests
-
-struct SopranoTextCleaningTests {
-
-    @Test func testTextCleaning() {
-        // Test number normalization
-        let text1 = "I have $100 and 50 cents."
-        let cleaned1 = cleanTextForSoprano(text1)
-        #expect(cleaned1.contains("one hundred dollars"), "Should expand dollar amounts")
-
-        // Test abbreviations
-        let text2 = "Dr. Smith went to the API conference."
-        let cleaned2 = cleanTextForSoprano(text2)
-        #expect(cleaned2.contains("doctor"), "Should expand Dr. to doctor")
-        #expect(cleaned2.contains("a p i"), "Should expand API")
-
-        // Test ordinals
-        let text3 = "This is the 1st and 2nd test."
-        let cleaned3 = cleanTextForSoprano(text3)
-        #expect(cleaned3.contains("first"), "Should expand 1st to first")
-        #expect(cleaned3.contains("second"), "Should expand 2nd to second")
-
-        print("\u{001B}[32mText cleaning tests passed!\u{001B}[0m")
-    }
-}
